refactor(security): split elevated allowFrom matcher internals
